* trilinos-develop: (411 commits)
  Disable tests in cuda rdc PT build already failing in non-rdc build (#4502)
  zoltan2:  shorten the test time
  Piro: slight change in how Final Time is read in.
  zoltan2:  fixed overflow in MJ
  zoltan2:  changed test name to include issue number
  zoltan2:  changed test name to include issue number
  zoltan2:  changed test name to include issue number
  zoltan2:  Removed unneeded complexity in test Error exposed itself with both int and double scalars; don't need to test both.
  Piro: putting in logic such that when running with the 'Albany Time Step Control Options' ParameterList, Tempus solve is killed upon reaching dt = Mininum Time Step, and encountering a failed solve.
  KokkosKernels: Updating patch of D2GC into MueLu Aggregation
  Automatic snapshot commit from seacas at ee0f81875
  Piro: Some cleanups to parsing of new 'Albany Time Step Control Options' parameter list.
  Panzer: add eblock multiplers to mass matrix tools
  Piro: adding an option in Piro::TempusSolver to take 'Albany Time Step Control Options' sublist, in place of the Tempus->Tempus Integrator->Time Step Control timelist.
  KokkosKernels: Patching in D2GC testing for kokkos-kernels package
  MueLu: Fix build warnings (unused typedefs)
  Ifpack2: Fix some build warnings
  Belos: Fix unused variable build warnings in a test
  Zoltan2: Fix #4813
  Zoltan2: Remove use of depr'c'd Tpetra::Crs{Graph,Matrix} ctors
  ...

+++ packages/stk/stk_balance/stk_balance/internal/balanceTypes.hpp
@@ -34,6 +34,9 @@
 #define BALANCETYPES_HPP_
 
 typedef int BalanceLocalNumber;
-typedef ssize_t BalanceGlobalNumber;
+
+//Zoltan2 insists that long long is supported at compile time so
+//it is safe for us to use it here.
+typedef long long BalanceGlobalNumber;
 
 #endif /* BALANCETYPES_HPP_ */
